Upgrade Detector with a faster, more efficient model (#1)
* Experiment with building a tflite model for text detection

This model based on yolo v8 I found seems like a good candidate! I can't
get Google's out-of-the-box ObjectDetector tool to work with it, but
running it directly with the InterpreterApi shows promise---we go from
~3s with comictextdetector to <1s!

Unfortunately, I have to figure out the format of the output, because it
is not working as promised when I call YOLO inference directly with the
source model in python....

* Fix: properly unpack the xyxy output from the yolo model!

Now we're cooking! Unfortunately the model seems to drop some of the
output that is normally emitted. Not sure what the deal is with that,
but it otherwise works fairly well? And it's still quite a bit faster
than the old model, even without quantization!

* Switch to the int8 model---it's < half the latency!!

We're now clocking about 300ms with pretty good accuracy. Anecdotally it
found the same boxes that the float32 model did, but in 30-50% of the
time! This represents an order of magnitude speed up from
comictextdetector---with a *lot* less CPU usage!!! Nice!

* Clean up the unused (unusable) vision task code

* Automatically fall back to the old detector if TFLite fails to load

* Add an inline mapRows util method

Not super important