Scaffold support for a 2-part OCR process in Android land
I *believe* we can potentially run the encoder just once, and reuse it
with new token IDs until we get the full output. If true, that could
potentially save us quite a bit of time.

Of course, the model doesn't actually work (it just outputs a full list
of NaN) so we'll have to uh make it work, first.

FWIW running the encoder in python *does* seem to produce non-garbage
values so... not totally sure what the deal is here. The decoder is
fully broken, however :(