Split OCR into two models for a huge speedup (#2)
* Prepare tools to train a custom OCR model

* Experiment with static quantization

Sadly seems to be slower than the dynamic quant we did before

* Fix CalibrationDataReader not providing labeled data

We just needed to tokenze the input! This could partially be what the PAD token is about: is there a way to do this as a one-shot somehow and get back a batch of 300 token IDs, instead of having to do it iteratively?

* Experiment with a manual reauthor

See: https://github.com/google-ai-edge/ai-edge-torch/tree/main/ai_edge_torch/generative#model-authoring-using-edge-generative-api

No idea if it's working, but we *do* have a separate encoder model in
tflite format using the original models weights, so... maybe?

* Add experimental reauthor of the decoder

There's a fair chance this is total garbage

* Experiment with trying to evaluate the converted models

* Scaffold support for a 2-part OCR process in Android land

I *believe* we can potentially run the encoder just once, and reuse it
with new token IDs until we get the full output. If true, that could
potentially save us quite a bit of time.

Of course, the model doesn't actually work (it just outputs a full list
of NaN) so we'll have to uh make it work, first.

FWIW running the encoder in python *does* seem to produce non-garbage
values so... not totally sure what the deal is here. The decoder is
fully broken, however :(

* Export just the encoder to tflite; fix android integration sorta

I mean it's still broken, but it doesn't crash?

* Fix: convert the decoder *only* to tflite!

* Configure decode export to have dynamic shape

* Experiment with using Onnx for the decoder

I *think* it's working? We're getting "no text" basically right now so
there's a very good chance the encoder is just not working right. But
the decode took only 6ms still, so there could be a path here? Perhaps
we can export the encoder in onnx format?

* Attempt to use an exported onnx version of the encoder

The shape seems more realistic and certainly the encoder output seems
like not garbage... Not sure what we're missing now

* Wire up processing

It's actually working totally fine! I just forgot that we get two token
2s at the beginning, and token 3 is when we stop.

This is *blazing* fast, and we haven't even quantized! This is more than
a factor of 5 speedup: ~110ms -> ~20ms per token! I suspect by
quantizing we can speed up the initial encoding as well, bringing the
average down even further.

* Refactor TfliteMangaOcr -> SplitPhaseMangaOcr

We may not actually end up using the tflite models at all, but this
refactor will make it easier to plug and play if we want to experiment!

* Clean up unused dependency that all of a sudden triggered build issue

* Quantize the split models

NOW THAT IS WHAT I'M TALKING ABOUT!! We're now down to ~10ms/token on
average, a full order of magnitude speedup!

* Add in sha hashes so we can download the new models

* Clean up the code to prepare for production use

* Clean up unused libs version

* Consolidate ModelPath classes

* Minor cleanup