Fix: failing to detect all text boxes (#3)
* Provide training data to YOLO for tflite int8 conversion

Doesn't seem to make a difference...

* Add an experimental LetterboxProcessor that's sometimes better

In our standard "detect" test case, the LetterboxProcessor catches all
of the boxes. However, it somehow performs *much worse* in most other
cases, missing many blocks of text---or somehow catching only half? It's
very strange.

* Save the last clipboard image to facilitate iteration

Much easier than having to constantly go back to the original app and
copy again

* Also handle an empty clipboard

* Fix letterbox processing with custom ResizeWithPadOp

The ResizeWithPadOrCrop was, well... cropping. So of course we lost some
text! Now with the aspect ratio-preserving resize, we've got quality
matching what we expect from desktop!

We might even be able to drop the non-class-zero results...?!

* Prefer only returning the "primary" class of results

This seems to match what YOLO is doing in python, and also seems to be
sufficient; the class 0 results in this new mode tend to be garbage
anyway.

* Add the missing sha hash for the new model