Further improve Detection using LETTERBOX with a high threshold
Turns out I got a bit ahead of myself with dropping the secondary
detection class entirely. There are some still pages that the text is
only detected as the secondary class. To account for the "garbage" in
this class (sometimes yomigana!) I've applied a somewhat high confidence
threshold to including the secondary class in the results.